readme update, clean up unnecessary code, refactor
